WIP - re-organized class structure
